requirements.txt
Django~=3.0.8
django-cors-headers~=3.4.0
djangorestframework~=3.11.0
django-extensions~=3.0.1
dj-rest-auth~=1.0.9
djangorestframework-simplejwt~=4.4.0
django-allauth~=0.42.0
django-mysql~=3.8.1
mysqlclient~=2.0.1
Pillow~=7.2.0